Update README.md
An algorithm to identify old people living alone when require help The final proposed code is written in C language & uses a series of logic based tests/ algorithms to test the seriousness of the situation and if it detects anything is fishy it alerts the primary contact/the emergency services in case primary contact is unable to respond on time. This project can add a group of values like security , health , care and assurance in the life of it’s elderly users. This programming project will not only serve it’s elder users facing health problems only but can also be used in a variety of other areas for eg. in ensuring the security and well being of children living alone. The program initially takes input the general information of the primary contact(the person who is going to be contacted initially in case of an emergency) and stores it in a struct data type. 

The proposed program uses the orientation/rotation sensors present in the existing android devices, it makes it extremely cost efficient and convenient to use. The proposed program takes input the initial orientation of the android device of the user. The device is either handheld e.g smartphones or wearable like a smart watch. Then it compares the initial orientation of the device during 96 consecutive times with the then orientation of the device and then it makes a decision based on the predefined steps. If it detects something is unusual it flashes a message to the user that the program is sending an SOS message to the primary contact. If the user finds that there is no need to send SOS then it can abort the SOS message by giving an input ‘N’ to the devices. If the device finds no input/another input it initiates the SOS. It waits for the response of primary contact and if it finds no response then goes ahead with contacting the emergency services itself.
This is just a console based model and can be implemented further in a working prototype.